+++ fabric-events-lifecycle-v0/build.gradle
@@ -3,6 +3,6 @@ version = getSubprojectVersion(project, "0.2.0")
 
 dependencies {
 	compile project(path: ':fabric-api-base', configuration: 'dev')
-	compile project(path: ':fabric-generic-events-v1', configuration: 'dev')
+	compile project(path: ':fabric-item-api-v1', configuration: 'dev')
 	compile project(path: ':fabric-lifecycle-events-v1', configuration: 'dev')
 }
