+++ fabric-events-lifecycle-v0/src/main/java/net/fabricmc/fabric/impl/event/lifecycle/client/LegacyClientEventInvokers.java
@@ -27,8 +27,9 @@ public class LegacyClientEventInvokers implements ClientModInitializer {
 	public void onInitializeClient() {
 		// Allows deprecated events to still be invoked by the newer implementations
 		ClientLifecycleEvents.CLIENT_TICK.register(client -> ClientTickCallback.EVENT.invoker().tick(client));
+		// Tick old events on ClientWorld
 		ClientLifecycleEvents.WORLD_TICK.register(world -> WorldTickCallback.EVENT.invoker().tick(world));
-		// This is part of generic events now.
-		net.fabricmc.fabric.api.client.event.generic.v1.item.ItemTooltipCallback.EVENT.register((stack, context, lines) -> ItemTooltipCallback.EVENT.invoker().getTooltip(stack, context, lines));
+		// This is part of item api now.
+		net.fabricmc.fabric.api.client.item.v1.ItemTooltipCallback.EVENT.register((stack, context, lines) -> ItemTooltipCallback.EVENT.invoker().getTooltip(stack, context, lines));
 	}
 }
