+++ fabric-events-lifecycle-v0/src/main/resources/fabric.mod.json
@@ -26,8 +26,8 @@
   "depends": {
     "fabricloader": ">=0.4.0",
     "fabric-api-base": "*",
-    "fabric-generic-events-v1": "*",
+    "fabric-item-api-v1": "*",
     "fabric-lifecycle-events-v1": "*"
   },
-  "description": "Legacy events for the game's lifecycle, superseded by fabric-lifecycle-events-v1 and fabric-generic-events-v1."
+  "description": "Legacy events for the game's lifecycle, superseded by fabric-lifecycle-events-v1 and fabric-item-api-v1."
 }
