--- fabric-generic-events-v1/src/main/resources/fabric-generic-events-v1.mixins.json
@@ -1,11 +1,0 @@
-{
-  "required": true,
-  "package": "net.fabricmc.fabric.mixin.event.generic",
-  "compatibilityLevel": "JAVA_8",
-  "client": [
-    "client.ItemStackMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
