--- fabric-generic-events-v1/src/main/resources/fabric.mod.json
@@ -1,26 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "fabric-generic-events-v1",
-  "name": "Fabric Generic Events (v1)",
-  "version": "${version}",
-  "environment": "*",
-  "license": "Apache-2.0",
-  "icon": "assets/fabric-generic-events-v1/icon.png",
-  "contact": {
-    "homepage": "https://fabricmc.net",
-    "irc": "irc://irc.esper.net:6667/fabric",
-    "issues": "https://github.com/FabricMC/fabric/issues",
-    "sources": "https://github.com/FabricMC/fabric"
-  },
-  "authors": [
-    "FabricMC"
-  ],
-  "mixins": [
-    "fabric-generic-events-v1.mixins.json"
-  ],
-  "depends": {
-    "fabricloader": ">=0.4.0",
-    "fabric-api-base": "*"
-  },
-  "description": "Events that don't explicitly belong to any other Fabric API modules."
-}
