--- fabric-generic-events-v1/src/testmod/resources/fabric.mod.json
@@ -1,16 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "fabric-generic-events-v1-testmod",
-  "name": "Fabric Generic Events (v1) Test Mod",
-  "version": "1.0.0",
-  "environment": "*",
-  "license": "Apache-2.0",
-  "depends": {
-    "fabric-generic-events-v1": "*"
-  },
-  "entrypoints": {
-    "client": [
-      "net.fabricmc.fabric.test.event.generic.TooltipTests"
-    ]
-  }
-}
