+++ fabric-item-api-v1/src/main/resources/assets/fabric-item-api-v1/icon.png
