+++ fabric-item-api-v1/src/testmod/java/net/fabricmc/fabric/test/item/client/TooltipTests.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright (c) 2016, 2017, 2018, 2019 FabricMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package net.fabricmc.fabric.test.item.client;
+
+import net.minecraft.text.LiteralText;
+import net.minecraft.util.Formatting;
+
+import net.fabricmc.api.ClientModInitializer;
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+import net.fabricmc.fabric.api.client.item.v1.ItemTooltipCallback;
+
+@Environment(EnvType.CLIENT)
+public class TooltipTests implements ClientModInitializer {
+	@Override
+	public void onInitializeClient() {
+		// Adds a tooltip to all items so testing can be verified easily.
+		ItemTooltipCallback.EVENT.register((stack, context, lines) -> {
+			lines.add(new LiteralText("Fancy Tooltips").formatted(Formatting.LIGHT_PURPLE));
+		});
+	}
+}
