+++ fabric-item-api-v1/src/testmod/resources/fabric.mod.json
@@ -1,0 +1,16 @@
+{
+  "schemaVersion": 1,
+  "id": "fabric-item-api-v1-testmod",
+  "name": "Fabric Item API (v1) Test Mod",
+  "version": "1.0.0",
+  "environment": "*",
+  "license": "Apache-2.0",
+  "depends": {
+    "fabric-item-api-v1": "*"
+  },
+  "entrypoints": {
+    "client": [
+      "net.fabricmc.fabric.test.item.client.TooltipTests"
+    ]
+  }
+}
