+++ settings.gradle
@@ -24,7 +24,7 @@ include 'fabric-content-registries-v0'
 include 'fabric-crash-report-info-v1'
 include 'fabric-events-interaction-v0'
 include 'fabric-events-lifecycle-v0'
-include 'fabric-generic-events-v1'
+include 'fabric-item-api-v1'
 include 'fabric-item-groups-v0'
 include 'fabric-keybindings-v0'
 include 'fabric-key-binding-api-v1'
